Add tolerance to some tests, so they work on Mac

+++ ssCSM_Linescan/validate.sh
@@ -19,19 +19,27 @@ rm -fv "$file.aux.xml"
 rm -fv "$gold.aux.xml"
 
 cmp_stats.sh $file $gold
-gdalinfo -stats $file | grep -v Files | grep -v -i tif > run.txt
-gdalinfo -stats $gold | grep -v Files | grep -v -i tif > gold.txt
+gdalinfo -stats $file | grep -v Files | grep -v -i tif | grep -i -v minimum > run.txt
+gdalinfo -stats $gold | grep -v Files | grep -v -i tif | grep -i -v minimum > gold.txt
 
 diff=$(diff run.txt gold.txt)
 cat run.txt
 
-rm -f run.txt gold.txt
+# Print the error and check the status
+../bin/max_err.pl run.txt gold.txt
+ans=$?
+if [ "$ans" -ne 0 ]; then
+	echo Validation failed
+	exit 1
+fi
 
-echo diff is $diff
-if [ "$diff" != "" ]; then
+ans=$(../bin/max_err.pl run.txt gold.txt 1e-5) # compare the error
+if [ "$ans" != "1" ]; then
     echo Validation failed
     exit 1
 fi
+
+rm -f run.txt gold.txt
 
 echo Validation succeeded
 exit 0
